When creating a user account try to derive the UID and GID from the homedir.
It's possible that when a new user is being created that the home
directory for that user already exists. When this happens account manager
will now try to find the UID of the user from that directory and reuse it.

This fixes potential problems if for some reason the new accounts that were
previously added into passwd/group were not preserved after reboot, but
the correspondinghome directory was.